fixes array domains not supported natively
Array domains do not have typelem set in pg_types and are not supported by default.

For example, this creates an array with a constraint ensuring that it has at least one element.

    CREATE DOMAIN public.page_quads AS polygon[]
       CONSTRAINT is_populated
       CHECK (COALESCE(array_length(VALUE, 1), 0) >= 1);

In this changeset, the bootstrap query used by the Type Server is updated again, so to support this use case without requiring a custom Postgrex extension.